MOR-92: Logo on the bar
