+++ Morphic.Bar/UI/BarImageControl.xaml
@@ -1,0 +1,47 @@
+﻿<ui:BarItemControl x:Class="Morphic.Bar.UI.BarImageControl"
+                         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+                         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                         xmlns:ui="clr-namespace:Morphic.Bar.UI"
+                         mc:Ignorable="d" d:DesignWidth="100" d:DesignHeight="70"
+                         MinHeight="87"
+                         d:DataContext="{d:DesignInstance ui:BarButtonControl, IsDesignTimeCreatable=True}"
+                         >
+    <ui:BarItemControl.Resources>
+        <!-- Colours based on the active theme -->
+        <SolidColorBrush x:Key="Background" Color="{Binding ActiveTheme.Background, FallbackValue=Yellow}"/>
+        <SolidColorBrush x:Key="Foreground" Color="{Binding ActiveTheme.TextColor, FallbackValue=Black}"/>
+        <SolidColorBrush x:Key="BorderColor" Color="{Binding ActiveTheme.BorderColor, FallbackValue=Transparent}"/>
+        <SolidColorBrush x:Key="BorderOuterColor" Color="{Binding ActiveTheme.BorderOuterColor, FallbackValue=Transparent}"/>
+        <!-- Convert a boolean to a visibility -->
+        <ui:Ternary x:Key="BoolVisibility" False="Collapse" True="Visible" />
+
+        <!-- Make the button style-less -->
+        <Style BasedOn="{StaticResource {x:Static ToolBar.ButtonStyleKey}}" TargetType="{x:Type Button}">
+            <Setter Property="Template">
+                <Setter.Value>
+                    <ControlTemplate TargetType="{x:Type Button}">
+                        <ContentPresenter />
+                    </ControlTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+
+    </ui:BarItemControl.Resources>
+    
+    <Button x:Name="Button"
+        Click="Button_OnClick"
+        HorizontalAlignment="Stretch" VerticalAlignment="Stretch"
+        HorizontalContentAlignment="Stretch" VerticalContentAlignment="Stretch"
+        >
+        <!-- The image -->
+        <Image
+            Source="{Binding BarItem.ImageSource}"
+            Visibility="{Binding BarItem.ShowIcon, Converter={StaticResource BoolVisibility}}"
+            Stretch="Uniform"
+            Margin="20"
+
+            />
+    </Button>
+</ui:BarItemControl>
